[#861] cleaner DI setup
