+++ android/src/main/kotlin/com/oztechan/ccc/android/Application.kt
@@ -13,7 +13,7 @@ import com.github.submob.logmob.initCrashlytics
 import com.github.submob.logmob.initLogger
 import com.oztechan.ccc.ad.initAds
 import com.oztechan.ccc.analytics.initAnalytics
-import com.oztechan.ccc.android.di.platformModule
+import com.oztechan.ccc.android.di.appModule
 import com.oztechan.ccc.client.di.initAndroid
 import mustafaozhan.github.com.mycurrencies.BuildConfig
 
@@ -40,7 +40,7 @@ class Application : Application() {
 
         initAndroid(
             context = this,
-            platformModule = platformModule
+            appModule = appModule
         )
 
         Thread.setDefaultUncaughtExceptionHandler(ANRWatchDogHandler())
