+++ backend/src/jvmMain/kotlin/com/oztechan/ccc/backend/di/modules/AppModule.kt
@@ -1,0 +1,5 @@
+package com.oztechan.ccc.backend.di.modules
+
+import com.oztechan.ccc.backend.di.modules.submodules.controllerModule
+
+internal val appModules = listOf(controllerModule)
