+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/di/module/ClientModules.kt
@@ -1,0 +1,11 @@
+package com.oztechan.ccc.client.di.module
+
+import com.oztechan.ccc.client.di.module.submodules.repositoryModule
+import com.oztechan.ccc.client.di.module.submodules.serviceModule
+import com.oztechan.ccc.client.di.module.submodules.viewModelModule
+
+val clientModules = listOf(
+    viewModelModule,
+    repositoryModule,
+    serviceModule,
+)
