+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/di/module/submodules/ServiceModule.kt
@@ -1,0 +1,11 @@
+package com.oztechan.ccc.client.di.module.submodules
+
+import com.oztechan.ccc.config.ConfigService
+import com.oztechan.ccc.config.ConfigServiceImpl
+import org.koin.core.module.dsl.bind
+import org.koin.core.module.dsl.singleOf
+import org.koin.dsl.module
+
+internal val serviceModule = module {
+    singleOf(::ConfigServiceImpl) { bind<ConfigService>() }
+}
