+++ client/src/commonMain/kotlin/com/oztechan/ccc/client/di/module/submodules/ViewModelModule.kt
@@ -1,10 +1,6 @@
-package com.oztechan.ccc.client.di.module
+package com.oztechan.ccc.client.di.module.submodules
 
-import com.oztechan.ccc.client.di.viewModelDefinition
-import com.oztechan.ccc.client.repository.background.BackgroundRepository
-import com.oztechan.ccc.client.repository.background.BackgroundRepositoryImpl
-import com.oztechan.ccc.client.repository.session.SessionRepository
-import com.oztechan.ccc.client.repository.session.SessionRepositoryImpl
+import com.oztechan.ccc.client.base.BaseViewModel
 import com.oztechan.ccc.client.viewmodel.adremove.AdRemoveViewModel
 import com.oztechan.ccc.client.viewmodel.calculator.CalculatorViewModel
 import com.oztechan.ccc.client.viewmodel.currencies.CurrenciesViewModel
@@ -12,13 +8,13 @@ import com.oztechan.ccc.client.viewmodel.main.MainViewModel
 import com.oztechan.ccc.client.viewmodel.selectcurrency.SelectCurrencyViewModel
 import com.oztechan.ccc.client.viewmodel.settings.SettingsViewModel
 import com.oztechan.ccc.client.viewmodel.watchers.WatchersViewModel
-import com.oztechan.ccc.config.ConfigService
-import com.oztechan.ccc.config.ConfigServiceImpl
-import org.koin.core.module.dsl.bind
-import org.koin.core.module.dsl.singleOf
+import org.koin.core.definition.Definition
+import org.koin.core.instance.InstanceFactory
+import org.koin.core.module.Module
+import org.koin.core.qualifier.Qualifier
 import org.koin.dsl.module
 
-var clientModule = module {
+internal val viewModelModule = module {
     viewModelDefinition { SettingsViewModel(get(), get(), get(), get(), get(), get()) }
     viewModelDefinition { MainViewModel(get(), get(), get()) }
     viewModelDefinition { CurrenciesViewModel(get(), get(), get()) }
@@ -26,8 +22,10 @@ var clientModule = module {
     viewModelDefinition { SelectCurrencyViewModel(get()) }
     viewModelDefinition { AdRemoveViewModel(get()) }
     viewModelDefinition { WatchersViewModel(get(), get()) }
-
-    singleOf(::ConfigServiceImpl) { bind<ConfigService>() }
-    singleOf(::SessionRepositoryImpl) { bind<SessionRepository>() }
-    singleOf(::BackgroundRepositoryImpl) { bind<BackgroundRepository>() }
 }
+
+expect inline fun <reified T : BaseViewModel> Module.viewModelDefinition(
+    qualifier: Qualifier? = null,
+    createdAtStart: Boolean = false,
+    noinline definition: Definition<T>
+): Pair<Module, InstanceFactory<T>>
