--- common/src/androidMain/kotlin/com/oztechan/ccc/common/di/modules/AndroidSettingsModule.kt
@@ -1,7 +1,0 @@
-package com.oztechan.ccc.common.di.modules
-
-import com.russhwolf.settings.AndroidSettings
-import com.russhwolf.settings.Settings
-import org.koin.core.scope.Scope
-
-actual fun Scope.provideSettings(): Settings = AndroidSettings(get())
