+++ common/src/androidMain/kotlin/com/oztechan/ccc/common/di/modules/submodules/AndroidDispatcherModule.kt
@@ -1,4 +1,4 @@
-package com.oztechan.ccc.common.di.modules
+package com.oztechan.ccc.common.di.modules.submodules
 
 import com.oztechan.ccc.common.di.DISPATCHER_DEFAULT
 import com.oztechan.ccc.common.di.DISPATCHER_IO
@@ -10,7 +10,7 @@ import org.koin.core.module.Module
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
-actual val dispatcherModule: Module = module {
+internal actual val dispatcherModule: Module = module {
     single<CoroutineDispatcher>(named(DISPATCHER_MAIN)) { Dispatchers.Main }
     single(named(DISPATCHER_IO)) { Dispatchers.IO }
     single(named(DISPATCHER_UNCONFINED)) { Dispatchers.Unconfined }
