+++ common/src/androidMain/kotlin/com/oztechan/ccc/common/di/modules/submodules/AndroidSettingsModule.kt
@@ -1,14 +1,18 @@
-package com.oztechan.ccc.client.di.module
+package com.oztechan.ccc.common.di.modules.submodules
 
 import android.content.Context
 import android.content.SharedPreferences
+import com.russhwolf.settings.AndroidSettings
+import com.russhwolf.settings.Settings
+import org.koin.core.module.dsl.bind
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.module
 
 private const val KEY_APPLICATION_PREFERENCES = "application_preferences"
 
-val androidModule = module {
+internal actual val settingsModule = module {
     singleOf(::provideSharedPreferences)
+    singleOf(::AndroidSettings) { bind<Settings>() }
 }
 
 private fun provideSharedPreferences(
