--- common/src/commonMain/kotlin/com/oztechan/ccc/common/di/modules/DispatcherModule.kt
@@ -1,5 +1,0 @@
-package com.oztechan.ccc.common.di.modules
-
-import org.koin.core.module.Module
-
-expect val dispatcherModule: Module
