--- common/src/commonMain/kotlin/com/oztechan/ccc/common/di/modules/SettingsModule.kt
@@ -1,11 +1,0 @@
-package com.oztechan.ccc.common.di.modules
-
-import com.russhwolf.settings.Settings
-import org.koin.core.scope.Scope
-import org.koin.dsl.module
-
-val settingsModule = module {
-    single { provideSettings() }
-}
-
-expect fun Scope.provideSettings(): Settings
