+++ common/src/commonMain/kotlin/com/oztechan/ccc/common/di/modules/submodules/DispatcherModule.kt
@@ -1,0 +1,5 @@
+package com.oztechan.ccc.common.di.modules.submodules
+
+import org.koin.core.module.Module
+
+internal expect val dispatcherModule: Module
