+++ common/src/commonMain/kotlin/com/oztechan/ccc/common/di/modules/submodules/ServiceModule.kt
@@ -1,4 +1,4 @@
-package com.oztechan.ccc.common.di.modules
+package com.oztechan.ccc.common.di.modules.submodules
 
 import com.oztechan.ccc.common.di.DISPATCHER_IO
 import com.oztechan.ccc.common.service.backend.BackendApiService
@@ -10,7 +10,7 @@ import com.oztechan.ccc.common.service.premium.PremiumApiServiceImpl
 import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
-val serviceModule = module {
+internal val serviceModule = module {
     single<FreeApiService> { FreeApiServiceImpl(get(), get(named(DISPATCHER_IO))) }
     single<BackendApiService> { BackendApiServiceImpl(get(), get(named(DISPATCHER_IO))) }
     single<PremiumApiService> { PremiumApiServiceImpl(get(), get(named(DISPATCHER_IO))) }
