--- common/src/jvmMain/kotlin/com/oztechan/ccc/common/di/modules/JVMSettingsModule.kt
@@ -1,8 +1,0 @@
-package com.oztechan.ccc.common.di.modules
-
-import com.russhwolf.settings.JvmPreferencesSettings
-import com.russhwolf.settings.Settings
-import org.koin.core.scope.Scope
-
-@Suppress("OPT_IN_USAGE")
-actual fun Scope.provideSettings(): Settings = JvmPreferencesSettings(get())
